uses second arduino for sd log, also tTime

+++ rocketArduino/globals.cpp
@@ -25,7 +25,7 @@ int countdownStart;
 
 String input = "";
 
-int tTime;
+long tTime;
 
 int burnHeight;
